Split material variable/run command et 1er refactor sur les noms/specialization de template
