+++ arcane/src/arcane/accelerator/RunCommandMaterialEnumerate.cc
@@ -1,0 +1,46 @@
+﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
+//-----------------------------------------------------------------------------
+// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// See the top-level COPYRIGHT file for details.
+// SPDX-License-Identifier: Apache-2.0
+//-----------------------------------------------------------------------------
+/*---------------------------------------------------------------------------*/
+/* RunCommRunCommandMaterialEnumerate.cc                       (C) 2000-2022 */
+/*                                                                           */
+/* Informations pour l'exécution d'une 'RunCommand'.                         */
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+#include "arcane/accelerator/RunCommRunCommandMaterialEnumerate.h"
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+namespace Arcane::Accelerator
+{
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+EnvCellRunCommand
+operator<<(RunCommand& command,const EnvCellVectorView& items)
+{
+  return EnvCellRunCommand(command,items);
+}
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+EnvCellRunCommand
+operator<<(RunCommand& command,IMeshEnvironment* env)
+{
+  return EnvCellRunCommand(command,env->envView());
+}
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+} // End namespace Arcane::Accelerator
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
